Initial mod stuff
The game loads and I've got some textures and stuff in place, but
unfortunately I can't get the block state information to propogate to
the renderer yet, so the server rakcs always face a particular direction
(despite the fact that if you right click on them right now they will
share their current facing direction.. so it's being stored and set
correctly in the world file)

Besides that, got some basic textures in (credit to Raven) and will be
improving on those as she has updates. Hopefully I didn't miss anything
but this is kinda a WIP commit.